requirements.txt
PyQt5
cairosvg
Pillow
